+++ lib/clevertap_plugin.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:ui';
+import 'dart:io';
 
 import 'package:clevertap_plugin/clevertap_plugin_web_wrapper.dart';
 import 'package:clevertap_plugin/src/types.dart';
@@ -43,10 +44,19 @@ class CleverTapPlugin {
       cleverTapPushClickedPayloadReceivedHandler;
   late CleverTapPushPermissionResponseReceivedHandler
       cleverTapPushPermissionResponseReceivedHandler;
+
   static List<CleverTapOnVariablesChangedHandler>
       cleverTapOnVariablesChangedHandlers = [];
-  static List<CleverTapOnValueChangedHandler> cleverTapOnValueChangedHandlers =
-      [];
+  static List<CleverTapOnOneTimeVariablesChangedHandler>
+      cleverTapOnOneTimeVariablesChangedHandlers = [];
+  static List<CleverTapOnValueChangedHandler> cleverTapOnValueChangedHandlers = [];
+  static List<CleverTapOnVariablesChangedAndNoDownloadsPendingHandler> cleverTapOnVariablesChangedAndNoDownloadsPendingHandlers = [];
+  static List<CleverTapOnceVariablesChangedAndNoDownloadsPendingHandler> cleverTapOnceVariablesChangedAndNoDownloadsPendingHandlers = [];
+  static List<CleverTapOnFileValueChangedHandler> cleverTapOnFileValueChangedHandlers = [];
+
+  late CleverTapCustomTemplatePresentHandler cleverTapCustomTemplatePresentHandler;
+  late CleverTapCustomFunctionPresentHandler cleverTapCustomFunctionPresentHandler;
+  late CleverTapCustomTemplateCloseHandler cleverTapCustomTemplateCloseHandler;
 
   static const MethodChannel _dartToNativeMethodChannel =
       const MethodChannel('clevertap_plugin/dart_to_native');
@@ -56,7 +66,10 @@ class CleverTapPlugin {
   static final CleverTapPlugin _clevertapPlugin =
       new CleverTapPlugin._internal();
 
-  factory CleverTapPlugin() => _clevertapPlugin;
+  factory CleverTapPlugin() {
+    print("CleverTapPlugin() called");
+    return _clevertapPlugin;
+  }
 
   static const libName = 'Flutter';
 
@@ -65,8 +78,7 @@ class CleverTapPlugin {
 
   CleverTapPlugin._internal() {
     /// Set the CleverTap Flutter library name and the current version for version tracking
-    _dartToNativeMethodChannel.invokeMethod(
-        'setLibrary', {'libName': libName, 'libVersion': libVersion});
+    _dartToNativeMethodChannel.invokeMethod('setLibrary', {'libName': libName, 'libVersion': libVersion});
     _nativeToDartMethodChannel.setMethodCallHandler(_platformCallHandler);
   }
 
@@ -149,6 +161,13 @@ class CleverTapPlugin {
           cleverTapOnVariablesChangedHandler(args.cast<String, dynamic>());
         });
         break;
+      case "onOneTimeVariablesChanged":
+        Map<dynamic, dynamic> args = call.arguments;
+        cleverTapOnOneTimeVariablesChangedHandlers
+            .forEach((cleverTapOnOneTimeVariablesChangedHandler) {
+          cleverTapOnOneTimeVariablesChangedHandler(args.cast<String, dynamic>());
+        });
+        break;
       case "onValueChanged":
         Map<dynamic, dynamic> args = call.arguments;
         cleverTapOnValueChangedHandlers
@@ -156,95 +175,170 @@ class CleverTapPlugin {
           cleverTapOnValueChangedHandler(args.cast<String, dynamic>());
         });
         break;
+      case "onVariablesChangedAndNoDownloadsPending":
+        Map<dynamic, dynamic> args = call.arguments;
+        cleverTapOnVariablesChangedAndNoDownloadsPendingHandlers
+            .forEach((cleverTapOnVariablesChangedAndNoDownloadsPendingHandler) {
+          cleverTapOnVariablesChangedAndNoDownloadsPendingHandler(args.cast<String, dynamic>());
+        });
+        break;      
+      case "onceVariablesChangedAndNoDownloadsPending":
+        Map<dynamic, dynamic> args = call.arguments;
+        cleverTapOnceVariablesChangedAndNoDownloadsPendingHandlers
+            .forEach((cleverTapOnceVariablesChangedAndNoDownloadsPendingHandler) {
+          cleverTapOnceVariablesChangedAndNoDownloadsPendingHandler(args.cast<String, dynamic>());
+        });
+        break;      
+      case "onFileValueChanged":
+        Map<dynamic, dynamic> args = call.arguments;
+        cleverTapOnFileValueChangedHandlers
+            .forEach((cleverTapOnFileValueChangedHandler) {
+          cleverTapOnFileValueChangedHandler(args.cast<String, dynamic>());
+        });
+        break;
+      case "customTemplatePresent":
+        String templateName = call.arguments;
+        cleverTapCustomTemplatePresentHandler(templateName);
+        break;
+      case "customFunctionPresent":
+        String templateName = call.arguments;
+        cleverTapCustomFunctionPresentHandler(templateName);
+        break;
+      case "customTemplateClose":
+        String templateName = call.arguments;
+        cleverTapCustomTemplateCloseHandler(templateName);
+        break;
       default:
         print('error');
         break;
     }
   }
 
+  /**
+   * Flushes any finished events on clevertap sdk side, maybe the client attaches
+   * listener at a later point and sdk has already tried to provide callback and
+   * failed
+   */
+  void invokeStartEmission(String name) {
+    if (Platform.isAndroid) {
+    _dartToNativeMethodChannel.invokeMethod('startEmission', name);
+    }
+  }
+  
+  void setCleverTapCustomTemplatePresentHandler(CleverTapCustomTemplatePresentHandler handler) {
+    invokeStartEmission('CleverTapCustomTemplatePresent');
+    cleverTapCustomTemplatePresentHandler = handler;
+  }
+  void setCleverTapCustomFunctionPresentHandler(CleverTapCustomTemplateCloseHandler handler) {
+    invokeStartEmission('CleverTapCustomFunctionPresent');
+    cleverTapCustomFunctionPresentHandler = handler;
+  }
+  void setCleverTapCustomTemplateCloseHandler(CleverTapCustomFunctionPresentHandler handler) {
+    invokeStartEmission('CleverTapCustomTemplateClose');
+    cleverTapCustomTemplateCloseHandler = handler;
+  }
+
   /// Define a method to handle inApp notification dismissed
-  void setCleverTapInAppNotificationDismissedHandler(
-          CleverTapInAppNotificationDismissedHandler handler) =>
-      cleverTapInAppNotificationDismissedHandler = handler;
+  void setCleverTapInAppNotificationDismissedHandler(CleverTapInAppNotificationDismissedHandler handler) {
+    invokeStartEmission('CleverTapInAppNotificationDismissed');
+    cleverTapInAppNotificationDismissedHandler = handler;
+  }
 
   /// Only for Android - Define a method to handle inApp notification shown
-  void setCleverTapInAppNotificationShowHandler(
-          CleverTapInAppNotificationShowHandler handler) =>
-      cleverTapInAppNotificationShowHandler = handler;
+  void setCleverTapInAppNotificationShowHandler(CleverTapInAppNotificationShowHandler handler) {
+    invokeStartEmission('CleverTapInAppNotificationShowed');
+    cleverTapInAppNotificationShowHandler = handler;
+  }
 
   /// Define a method to handle inApp notification button clicked
-  void setCleverTapInAppNotificationButtonClickedHandler(
-          CleverTapInAppNotificationButtonClickedHandler handler) =>
-      cleverTapInAppNotificationButtonClickedHandler = handler;
+  void setCleverTapInAppNotificationButtonClickedHandler(CleverTapInAppNotificationButtonClickedHandler handler) {
+    invokeStartEmission('CleverTapInAppNotificationButtonTapped');
+    cleverTapInAppNotificationButtonClickedHandler = handler;
+  }
 
   /// Define a method to handle profile initialization
-  void setCleverTapProfileDidInitializeHandler(
-          CleverTapProfileDidInitializeHandler handler) =>
-      cleverTapProfileDidInitializeHandler = handler;
+  void setCleverTapProfileDidInitializeHandler(CleverTapProfileDidInitializeHandler handler) {
+    invokeStartEmission('CleverTapProfileDidInitialize');
+    cleverTapProfileDidInitializeHandler = handler;
+  }
 
   /// Define a method to handle profile sync
-  void setCleverTapProfileSyncHandler(CleverTapProfileSyncHandler handler) =>
-      cleverTapProfileSyncHandler = handler;
+  void setCleverTapProfileSyncHandler(CleverTapProfileSyncHandler handler) {
+    invokeStartEmission('CleverTapProfileSync');
+    cleverTapProfileSyncHandler = handler;
+  }
 
   /// Define a method to handle inbox initialization
-  void setCleverTapInboxDidInitializeHandler(
-          CleverTapInboxDidInitializeHandler handler) =>
-      cleverTapInboxDidInitializeHandler = handler;
+  void setCleverTapInboxDidInitializeHandler(CleverTapInboxDidInitializeHandler handler) {
+    invokeStartEmission('CleverTapInboxDidInitialize');
+    cleverTapInboxDidInitializeHandler = handler;
+  }
 
   /// Define a method to handle inbox update
-  void setCleverTapInboxMessagesDidUpdateHandler(
-          CleverTapInboxMessagesDidUpdateHandler handler) =>
-      cleverTapInboxMessagesDidUpdateHandler = handler;
+  void setCleverTapInboxMessagesDidUpdateHandler(CleverTapInboxMessagesDidUpdateHandler handler) {
+    invokeStartEmission('CleverTapInboxMessagesDidUpdate');
+    cleverTapInboxMessagesDidUpdateHandler = handler;
+  }
 
   /// Define a method to handle inbox notification button clicked
-  void setCleverTapInboxNotificationButtonClickedHandler(
-          CleverTapInboxNotificationButtonClickedHandler handler) =>
-      cleverTapInboxNotificationButtonClickedHandler = handler;
+  void setCleverTapInboxNotificationButtonClickedHandler(CleverTapInboxNotificationButtonClickedHandler handler) {
+    invokeStartEmission('CleverTapInboxMessageButtonTapped');
+    cleverTapInboxNotificationButtonClickedHandler = handler;
+  }
 
   /// Define a method to handle inbox notification message clicked
-  void setCleverTapInboxNotificationMessageClickedHandler(
-          CleverTapInboxNotificationMessageClickedHandler handler) =>
-      cleverTapInboxNotificationMessageClickedHandler = handler;
+  void setCleverTapInboxNotificationMessageClickedHandler(CleverTapInboxNotificationMessageClickedHandler handler) {
+    invokeStartEmission('CleverTapInboxMessageTapped');
+    cleverTapInboxNotificationMessageClickedHandler = handler;
+  }
 
   /// Define a method to handle Native Display Unit updates
-  void setCleverTapDisplayUnitsLoadedHandler(
-          CleverTapDisplayUnitsLoadedHandler handler) =>
-      cleverTapDisplayUnitsLoadedHandler = handler;
+  void setCleverTapDisplayUnitsLoadedHandler(CleverTapDisplayUnitsLoadedHandler handler) {
+    invokeStartEmission('CleverTapDisplayUnitsLoaded');
+    cleverTapDisplayUnitsLoadedHandler = handler;
+  }
 
   /// Define a method to handle Feature Flag updates
-  void setCleverTapFeatureFlagUpdatedHandler(
-          CleverTapFeatureFlagUpdatedHandler handler) =>
-      cleverTapFeatureFlagUpdatedHandler = handler;
+  void setCleverTapFeatureFlagUpdatedHandler(CleverTapFeatureFlagUpdatedHandler handler) {
+    invokeStartEmission('CleverTapFeatureFlagsDidUpdate');
+    cleverTapFeatureFlagUpdatedHandler = handler;
+  }
 
   /// Define a method to handle Product config initialization
-  void setCleverTapProductConfigInitializedHandler(
-          CleverTapProductConfigInitializedHandler handler) =>
-      cleverTapProductConfigInitializedHandler = handler;
+  void setCleverTapProductConfigInitializedHandler(CleverTapProductConfigInitializedHandler handler) {
+    invokeStartEmission('CleverTapProductConfigDidInitialize');
+    cleverTapProductConfigInitializedHandler = handler;
+  }
 
   /// Define a method to handle Product config fetch updates
-  void setCleverTapProductConfigFetchedHandler(
-          CleverTapProductConfigFetchedHandler handler) =>
-      cleverTapProductConfigFetchedHandler = handler;
+  void setCleverTapProductConfigFetchedHandler(CleverTapProductConfigFetchedHandler handler) {
+    invokeStartEmission('CleverTapProductConfigDidFetch');
+    cleverTapProductConfigFetchedHandler = handler;
+  }
 
   /// Define a method to handle Product config activation updates
-  void setCleverTapProductConfigActivatedHandler(
-          CleverTapProductConfigActivatedHandler handler) =>
-      cleverTapProductConfigActivatedHandler = handler;
+  void setCleverTapProductConfigActivatedHandler(CleverTapProductConfigActivatedHandler handler) {
+    invokeStartEmission('CleverTapProductConfigDidActivate');
+    cleverTapProductConfigActivatedHandler = handler;
+  }
 
   /// Define a method to handle Push Amplification payload
-  void setCleverTapPushAmpPayloadReceivedHandler(
-          CleverTapPushAmpPayloadReceivedHandler handler) =>
-      cleverTapPushAmpPayloadReceivedHandler = handler;
+  void setCleverTapPushAmpPayloadReceivedHandler(CleverTapPushAmpPayloadReceivedHandler handler) {
+    invokeStartEmission('CleverTapPushAmpPayloadReceived');
+    cleverTapPushAmpPayloadReceivedHandler = handler;
+  }
 
   /// Define a method to handle Push Clicked payload
-  void setCleverTapPushClickedPayloadReceivedHandler(
-          CleverTapPushClickedPayloadReceivedHandler handler) =>
-      cleverTapPushClickedPayloadReceivedHandler = handler;
+  void setCleverTapPushClickedPayloadReceivedHandler(CleverTapPushClickedPayloadReceivedHandler handler) {
+    invokeStartEmission('CleverTapPushNotificationClicked');
+    cleverTapPushClickedPayloadReceivedHandler = handler;
+  }
 
   /// Define a method to handle Push permission response
-  void setCleverTapPushPermissionResponseReceivedHandler(
-          CleverTapPushPermissionResponseReceivedHandler handler) =>
-      cleverTapPushPermissionResponseReceivedHandler = handler;
+  void setCleverTapPushPermissionResponseReceivedHandler(CleverTapPushPermissionResponseReceivedHandler handler) {
+    invokeStartEmission('CleverTapPushPermissionResponseReceived');
+    cleverTapPushPermissionResponseReceivedHandler = handler;
+  }
 
   /// Set a message handler function which is called when the app is in the
   /// terminated or killed state.
@@ -1129,6 +1223,13 @@ class CleverTapPlugin {
         .invokeMethod('defineVariables', {'variables': variables});
   }
 
+  ///Create File variable.
+  /// * @param {object} variables The JSON Object specifying the varibles to be created.
+  static Future<void> defineFileVariable(String fileVariable) async {
+    return await _dartToNativeMethodChannel
+        .invokeMethod('defineFileVariable', {'fileVariable': fileVariable});
+  }
+
   ///Get all variables via a JSON object.
   static Future<Map<Object?, Object?>> getVariables() async {
     return await _dartToNativeMethodChannel.invokeMethod('getVariables', {});
@@ -1150,6 +1251,11 @@ class CleverTapPlugin {
     }
   }
 
+  static void onOneTimeVariablesChanged(CleverTapOnOneTimeVariablesChangedHandler handler) {
+      cleverTapOnOneTimeVariablesChangedHandlers.add(handler);
+      _dartToNativeMethodChannel.invokeMethod('onOneTimeVariablesChanged', {});
+  }
+
   static void onValueChanged(
       String name, CleverTapOnValueChangedHandler handler) {
     if (!kIsWeb) {
@@ -1158,6 +1264,21 @@ class CleverTapPlugin {
     } else {
       CleverTapPluginWeb.onValueChanged(name, handler);
     }
+  }
+
+  static void onVariablesChangedAndNoDownloadsPending(CleverTapOnVariablesChangedAndNoDownloadsPendingHandler handler) {
+      cleverTapOnVariablesChangedAndNoDownloadsPendingHandlers.add(handler);
+      _dartToNativeMethodChannel.invokeMethod('onVariablesChangedAndNoDownloadsPending', {});
+  }
+
+  static void onceVariablesChangedAndNoDownloadsPending(CleverTapOnceVariablesChangedAndNoDownloadsPendingHandler handler) {
+      cleverTapOnceVariablesChangedAndNoDownloadsPendingHandlers.add(handler);
+      _dartToNativeMethodChannel.invokeMethod('onceVariablesChangedAndNoDownloadsPending', {});
+  }
+
+  static void onFileValueChanged(String name, CleverTapOnFileValueChangedHandler handler) {
+      cleverTapOnFileValueChangedHandlers.add(handler);
+      _dartToNativeMethodChannel.invokeMethod('onFileValueChanged', {'name': name});
   }
 
   ///Sets the user locale.
@@ -1175,5 +1296,66 @@ class CleverTapPlugin {
   static Future<void> clearInAppResources(bool expiredOnly) async {
     return await _dartToNativeMethodChannel.invokeMethod(
         'clearInAppResources', expiredOnly);
+  }
+
+
+  /**
+     * Uploads Custom in-app templates and app functions to the server.
+     * Requires Development/Debug build/configuration.
+     */
+  
+  static Future<void> syncCustomTemplates() async {
+    return await _dartToNativeMethodChannel
+        .invokeMethod('syncCustomTemplates', {});
+  }
+
+  static Future<void> syncCustomTemplatesInProd(bool isProduction) async {
+    return await _dartToNativeMethodChannel
+        .invokeMethod('syncCustomTemplatesInProd', {'isProduction': isProduction});
+  }
+
+  static Future<void> customTemplateSetDismissed(String templateName) async {
+    return await _dartToNativeMethodChannel
+        .invokeMethod('customTemplateSetDismissed', templateName);
+  }
+
+  static Future<void> customTemplateSetPresented(String templateName) async {
+    return await _dartToNativeMethodChannel
+        .invokeMethod('customTemplateSetPresented', templateName);
+  }
+
+  static Future<void> customTemplateRunAction(String templateName, String argName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateRunAction',
+        {'templateName': templateName, 'argName': argName});
+  }
+
+  static Future<String?> customTemplateGetStringArg(String templateName, String argName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateGetStringArg',
+        {'templateName': templateName, 'argName': argName});
+  }
+
+  static Future<num?> customTemplateGetNumberArg(String templateName, String argName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateGetNumberArg',
+        {'templateName': templateName, 'argName': argName});
+  }
+
+  static Future<bool?> customTemplateGetBooleanArg(String templateName, String argName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateGetBooleanArg',
+        {'templateName': templateName, 'argName': argName});
+  }
+
+  static Future<String?> customTemplateGetFileArg(String templateName, String argName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateGetFileArg',
+        {'templateName': templateName, 'argName': argName});
+  }
+
+  static Future<Object?> customTemplateGetObjectArg(String templateName, String argName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateGetObjectArg',
+        {'templateName': templateName, 'argName': argName});
+  }
+
+  static Future<String?> customTemplateContextToString(String templateName) async {
+    return await _dartToNativeMethodChannel.invokeMethod('customTemplateContextToString',
+        {'templateName': templateName});
   }
 }